requirements.txt
anyio==4.9.0
certifi==2025.4.26
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
joblib==1.5.1
numpy==1.26.4

pillow==11.2.1
scikit-learn==1.6.1
scipy==1.15.3
sniffio==1.3.1
threadpoolctl==3.6.0
typing_extensions==4.14.0


# requirements.txt

# Core FastAPI
fastapi
uvicorn[standard]

# Data Validation and Settings
pydantic
pydantic-settings
python-dotenv
email-validator

# AI/LLM
google-generativeai
httpx
Pillow

# Database ORM and Migrations
sqlalchemy[asyncio]
alembic
greenlet

# Database Drivers
aiosqlite
psycopg2-binary
asyncpg

# Security and Authentication
passlib[bcrypt]
python-jose[cryptography]

# --- Image Analysis (for Apple Silicon M1/M2/M3 Macs) ---
deepface
# For Apple Silicon, install tensorflow-macos provided by Apple
tensorflow-macos
# tensorflow-metal is needed for GPU acceleration
tensorflow-metal
opencv-python
mediapipe
tf-keras
# NOTE: We are removing specific versions to let pip resolve the best
# compatible versions for Apple Silicon and Python 3.10.
deepface==0.0.92
tensorflow==2.15.0
tf-keras==2.15.0
opencv-python==4.9.0.80
numpy==1.26.4
geopy
beautifulsoup4
selenium
webdriver-manager
selenium-stealth
#vector databases 
chromadb
